create ColorSystem class and colors.json file
update DataLoader to load color info from colors.json and update core/index.js to include references to ColorSystem

add ColorSystem initAsync to Context.js, create getColorScheme method on ColorSystem class, import getColorScheme in styles.js

start color refactor in styles.js

bind getColorScheme method to 'this' and update constructor

refactor styles.js into STYLES class and update import statements to reflect changes

remove initAsync() call in Context.js file

change class name to 'Styles', update imports, change DEFAULT to 'default'

update styleMatch method, change all vars to static vars, fix formatting

add script to build_data.js to minify colors.json

remove static colors.json import, refactor getColorScheme method

start Styles refactor to StyleSystem, move file to /modules/core, update import statements

add StyleSystem reference to available systems set

add styles system dependency to MapSystem

remove StyleSystem import, update style vars to use context.systems

refactor StyleSystem

initialize colorData as class var, move getDataAsync function to startAsync

remove StyleSystem import, set style var with context

delete styles.js

update MapSystem.Test to include colors and styles
