+++ data/colors.json
@@ -1,0 +1,38 @@
+{
+    "red": {
+      "fill": { "color": "0xe06e5f", "alpha": 0.3 }  
+    },
+    "green": {
+      "fill": { "color": "0x8cd05f", "alpha": 0.3 }  
+    },
+    "blue": {
+      "fill": { "color": "0x77d4de", "alpha": 0.3 }   
+    },
+    "yellow": {
+      "fill": { "color": "0xffff94", "alpha": 0.25 }  
+    },
+    "gold": {
+      "fill": { "color": "0xc4be19", "alpha": 0.3 }  
+    },
+    "orange": {
+      "fill": { "color": "0xd6881a", "alpha": 0.3 }  
+    },
+    "pink": {
+      "fill": { "color": "0xe3a4f5", "alpha": 0.3 }   
+    },
+    "teal": {
+      "fill": { "color": "0x99e1aa", "alpha": 0.3 }   
+    },
+    "lightgreen": {
+      "fill": { "color": "0xbee83f", "alpha": 0.3 }  
+    },
+    "tan": {
+      "fill": { "color": "0xf5dcba", "alpha": 0.3 }   
+    },
+    "darkgray": {
+      "fill": { "color": "0x8c8c8c", "alpha": 0.5 }   
+    },
+    "lightgray": {
+      "fill": { "color": "0xaaaaaa", "alpha": 0.3 }   
+    }
+ }
