+++ modules/core/DataLoaderSystem.js
@@ -44,6 +44,7 @@ export class DataLoaderSystem extends AbstractSystem {
     fileMap.set('shortcuts', 'data/shortcuts.min.json');
     fileMap.set('territory_languages', 'data/territory_languages.min.json');
     fileMap.set('wmf_sitematrix', 'https://cdn.jsdelivr.net/npm/wmf-sitematrix@0.1/wikipedia.min.json');
+    fileMap.set('colors', 'data/colors.min.json');
 
     this.fileMap = fileMap;
     this._cachedData = {};
@@ -74,6 +75,7 @@ export class DataLoaderSystem extends AbstractSystem {
       c.shortcuts = [];
       c.territory_languages = {};
       c.wmf_sitematrix = [ ['English','English','en'], ['German', 'Deutsch', 'de'] ];
+      c.colors = {};
     }
   }
